Done for the night. Still more to go....:( :( Meeting up tomorrow.

+++ AsianDomination/DummyClasses/SequenceDiagramPack/driver.java
@@ -4,7 +4,8 @@ public class driver {
 	public static void main(String[] args) {
 		Sale s = new Sale();
 		Register r = new Register(s);
-		int a = 10;
+		double a = 10.0;
 		r.checkout(a);
+//		r.out(true, 10.0, 2);
 	}
 }
